requirements.txt
pytest
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
allure-pytest
